+++ src/main/java/com/needle/FsoFso/product/services/ProductService.java
@@ -13,7 +13,7 @@ public interface ProductService {
 	
 	int getAllProduct();
 	
-	ProductDto getProductById(int id);
+	ProductDto getProductById(long id);
 	
 	void addCart(CartDto cart);
 }
